Convert the reprepro class to parametrized format
We remove the deprecated dynamic lookup for variables and now have them as
class parameters.

The reprepro::cron and reprepro::inoticoming classes have been integrated
in the reprepro class as two boolean arguments. As these are not mutually
exclusive, it is cleaner that way.

We add a test manifest for the reprepro class along the way.