Revert "fix permissions recursively in directories"
This makes it impossible to apply a catalog on a VM with 512MB of RAM (that has
1.5k files in pool/, and 10k files in dists/).

Apparently, the change is only needed when "we change the owner of
a repository", which looks rare enough to me not to deserve such implications.
A better way to support this usecase should be found.

This reverts commit 4f8da394948ba7544a756930d363ea2d7aca291e.

Conflicts:
	manifests/repository.pp